Extract session-independent exchange state capture and restore
